.github: pin ubuntu-latest as ubuntu-22.04
The build is via Docker, so there shouldn't be much scope for an
automatic bump from e.g. Ubuntu 22.04 to Ubuntu 24.04 to break the
build. However, pinning here is still good for reproducibility,
especially in the period where `ubuntu-latest` may run on either
Ubuntu 22.04 or 24.04. The Ubuntu 22.04 transition took 8 weeks [1]:

    Now Ubuntu-22.04 is ready to be the default version for the
    `ubuntu-latest` label in GitHub Actions workflows. This change will be
    rolled out over a period of 8 weeks beginning on October 1, 2022.

This is also consistent with the arm64 build job, which is pinned to
Ubuntu 22.04.

Currently, `ubuntu-latest` runs on Ubuntu 22.04, and Ubuntu 24.04 is in
beta [2][3].

[1] https://github.blog/changelog/2022-11-09-github-actions-ubuntu-latest-workflows-will-use-ubuntu-22-04/
[2] https://github.com/actions/runner-images/blob/380b0afd7e43/README.md#available-images
[3] https://github.blog/changelog/2024-05-14-github-hosted-runners-public-beta-of-ubuntu-24-04-is-now-available/