.github, Dockerfile.ubuntu: use bundled Nimble version by default
Our workflow set the defaults such that Nim 2.0.0 was installed, plus
the (still-)latest Nimble tag (0.14.2, 2023-02-20). However, that Nimble
version is behind the one that was bundled with Nim 2.0.0 [1][2][3].

Change the release workflow default such that only the release-defined
version of Nimble is installed.

Nim 2.0.8 (2024-07-03) bundles a Nimble version from 2024-06-27 [4][5],
and staying with Nimble 0.14.2 would likely produce unexpected behavior,
bugs, or missing functionality.

We could instead write the release-defined Nimble hash as the default
NIMBLE_VERSION, but writing something like `bundled` gives a single
source of truth and seems clearer.

Later, we could consider removing the ability to specify a separate
Nimble version. But for now, leave that ability, and preserve the Nim
1.6.12 and Nimble 0.14.0 pairing that was present in the Dockerfile.

Background:

Dockerfile.ubuntu runs `./koch tools` to build Nimble (along with
various other tools, like nimsuggest and testament). The Nimble version
that's built in this step is defined by koch itself; for example, for
Nim 2.0.0 it builds Nimble from commit 168416290e49 [1]. This is what
defines the Nimble version that's included in the prebuilt Nim releases.

However, that Dockerfile later installs another Nimble version, defined
by `NIMBLE_VERSION`, and manipulates PATH to make that version take
precedence.

I assume that was done in order to make Nim 1.6.12 use a more recent
version than that bundled. But for newer Nim releases, it's better if
running a `nimble` command uses the Nimble version defined by the Nim
release we're using.

[1] https://github.com/nim-lang/Nim/blob/v2.0.0/koch.nim#L14
[2] https://github.com/nim-lang/nimble/compare/v0.14.2...168416290e49
[3] https://github.com/nim-lang/nimble/releases/tag/v0.14.2
[4] https://github.com/nim-lang/Nim/blob/v2.0.8/koch.nim#L14
[5] https://github.com/nim-lang/nimble/commit/be2f1309b35a